teams_bot/requirements.txt
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botbuilder-integration-aiohttp==4.17.0
botframework-connector==4.17.0
aiohttp>=3.9.4
azure-data-tables>=12.4.0
azure-identity>=1.15.0
